+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All breaking changes are marked with [BC] and potentially require API consumer changes after updating to the respective version.
 
+## [Unreleased]
+### Fixed
+- Fix wrapping cursor outside of visible window.
+
 ## [0.2.0] - 2019-07-20
 ### Added
 - Add Default variant to base::Color enum. [BC]